typelanguage: start generating actual Rust (incomplete)
