+++ Cargo.toml
@@ -3,9 +3,13 @@ name = "re-utilities"
 version = "0.1.0"
 edition = "2021"
 
+# todo: define a feature for typelanguage
+
 [dependencies]
 anyhow = "1.0.56"
 glob = "0.3.0"
+itertools = "0.10.3"
+quote = "1.0.16"
 syn = "1.0.89"
 patternscan = "1.2.0"
 proc-macro2 = { version = "1.0.36", features = ["span-locations"] }
