+++ src/type_language/mod.rs
@@ -1,3 +1,4 @@
 pub mod driver;
 pub mod grammar;
 pub mod parser;
+pub mod semantic_analysis;
